feat: ⬆️ examples: upgrade yadm, update dotfiles phase, modularize phases, and more
- Examples:
    - Upgrade yadm
    - Update dotfiles phase
    - Modularize phases
 - Development sandbox:
    - Upgrade Ubuntu image
    - Fix bashrc
 - GitHub actions: Merge e2e tests, remove redundancies
 - Update eslint, prettier configs
 - Rm old reference files
