--- .github/workflows/e2e-test-bootstrap-apply-master.yml
@@ -1,36 +1,0 @@
-name: "E2E tests: Bootstrap / apply (master)"
-
-on:
-  push:
-    branches:
-      - master
-
-defaults:
-  run:
-    shell: bash
-
-jobs:
-  master-defaults:
-    strategy:
-      fail-fast: true
-      matrix:
-        os: [macos-latest, macos-11.0, ubuntu-latest, ubuntu-20.04]
-
-    runs-on: ${{ matrix.os }}
-
-    steps:
-      - name: Bootstrap Akinizer from script
-        run: curl -o- https://raw.githubusercontent.com/robatron/akinizer/master/bootstrap.sh | bash
-
-      - name: Apply Akinizer configs
-        run: |
-          cd $HOME/opt/akinizer/examples
-          gulp
-
-      - name: Re-bootstrap non-destructively
-        run: curl -o- https://raw.githubusercontent.com/robatron/akinizer/master/bootstrap.sh | bash
-
-      - name: Re-apply non-destructively
-        run: |
-          cd $HOME/opt/akinizer/examples
-          gulp
