--- ref/.Brewfile
@@ -1,15 +1,0 @@
-# target directory for casks
-cask_args appdir: '/Applications'
-
-# install packages
-tap 'caskroom/homebrew-cask' || true
-brew 'coreutils'
-brew 'cowsay'
-brew 'fortune'
-brew 'git'
-brew 'gpg'
-brew 'neovim'
-brew 'pyenv'
-brew 'reattach-to-user-namespace'
-brew 'tmux'
-brew 'zplug'
