--- ref/README.md
@@ -1,1 +1,0 @@
-Old bootstrapping files for reference
